First commit to CLI functionality; stubs
Code is stubs, but MARC record searching works fine.  Many FIXME tags.
Partial I18N, but on a CLI, this is probably silly; we can probably just
call everything with language set to CFG_SITE_LANG and call it a day.  The
chance that a user will want the CLI output help in Chinese seems vanishingly
small.

Going home for the weekend.