BibEdit: Author Special exploratory
Creates a new URL: /record/editauthors, which works substantially like the
existing BibEdit interface, with the following modifications:
- only author and affiliation fields are editable
- the first author field is opened for editing on page load
- editable fields are bolded, and non-editable fields are plain
- hotkey behavior is modified slightly: return accepts this field, and expands
  the next one which may be worked on

BibEdit Special Authors 1st commit

Oops; forgot some files.

Added author engine

Committing as much as I know how to do up to this point.

Today's work

First pass at cosmetic javascript fixes.

When a record is loaded in the author editor (really a minimally modified
BibEdit), do minor tweaking to the CSS to make the edit-worthy parts
stand out, and also send a click event to the first author (so that it
expands into an edit box with the cursor homed in it.)

Specialization of hotkey behavior.

Author special mode moves to the next relevant subfield and expands it when
return is pressed in a subfield.  When the last relevant subfield is reached,
jumps back to the top (100a).

Hotkey fix implemented as a fairly minimal patch over top of the existing
bibedit_keys.js infrastructure.  This may be a good way to do the rest of
the changes necessary to support special modes, once BibEdit is suitably
refactored.

TODO: validation/confidence wrt external data sources/sinks
TODO: patch BibEdit so it generates not CSS validation complaints,
      and so that code is repeated as little as possible - even if
      this means the introduction of *gasp* objects

Tab/Space issues in bibeditauthors_engine.js

In many places in bibeditauthors_engine.js, tabs are intermixed along with the
spaces.  Also, there are a number of places where the indention level seems
arbitrary; some of these have been corrected to (more) semantic sense.

Whitespace corrections only