BibRecord: SKIPPED tests, scrutability
Re: SKIPPED tests in bibrecord_tests:

Unit tests should fail if and only if the tested functionality is both
expected to work, and does not.  A SKIPPED test is not the same thing as
a FAILED test, and using TestCase.fail() is inappropriate.

A correctly installed and configured system working from a stable code
tree should never have failing unit tests.

This patch causes a warning to be emitted, but lets the skipped tests
count as a pass.

Re: scrutability of bibrecord:

I realized I was violating DRY in my SKIPPED test update.  Fixed.

I rewrote record_get_field_values to be shorter and easier to read.