asmv2: Initial checkin
Rebasing against master and squashing 34 commits.

A-C-S-a invokes smart-replacement.  Try putting '08544' into an
affiliation box, hit the magic key sequence, and any right-minded
KB will give you 'Princeton U.'

There are a few things that keep this from being really finished, such as
the handcrafted test KB, or the fact that it's called "JoeTest".  But it's
working as a prototype, so I feel ok moving on to the next feature.

Displays column numbers on columns and labels.  Puts the number portion in
the column header into its own <span>, so it can be styled separately.
FIXME: does styling inline to show this off.

The other half of column numbering.  You can input the small integers used
to tag the columns in the affiliation boxes, and they'll be transparently
derefenced into full institution names by the onChange handler.

changes to pass kwalitee checks