refextract: making cli defaults sensible
The cli_opts dictionary is now being initialized so that it's values are
reasonable defaults even when they're not explicitly set.  This means that
code which uses refextract as a library only has to specify the options that
differ from these default values (all 0's).

FIXME: refextract should be using optparse instead.  Seriously.