editauthor: bug fix introducing new bug
When autosuggest is invoked, the pre-invocation value is added as a check
column.  Now, the post-invocation value is added also (which is correct).

To fix, we need to either autosuggest on everything in the .change() handler,
or break up the update process into a set of layers, and call into that stack
of layers from within the autosuggest key handler.